refactor user type definition
