+++ runarion-laravel/resources/js/types/index.d.ts
@@ -1,10 +1,4 @@
-export interface User {
-    id: number;
-    name: string;
-    email: string;
-    settings: Record<string, string | number | boolean>;
-    email_verified_at?: string;
-}
+import type { User } from "@/types/user";
 
 export type PageProps<
     T extends Record<string, unknown> = Record<string, unknown>
